Modularize maker condition builder
